requirements.txt
aiogram
supabase
yt-dlp
python-dotenv
asyncio
aiohttp
